requirements.txt
fastapi
uvicorn
python-multipart
spacy
scikit-learn
pydantic-settings
pypdf
python-docx
pytest
